backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
pyspellchecker==0.8.1
google-generativeai==0.7.2
python-docx==1.1.2
reportlab==4.2.2
